backend/requirements.txt
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
Flask-JWT-Extended==4.6.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.6
importlib_resources==6.1.2
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mysqlclient==2.2.4
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
SQLAlchemy==2.0.27
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
